requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auto_gptq==0.7.1
autoawq==0.2.7.post3
blessed==1.20.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
DataProperty==1.0.1
datasets==3.2.0
debugpy==1.8.7
decorator==5.1.1
deepspeed==0.16.3
dill==0.3.7
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2023.10.0
gekko==1.2.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
-e git+https://github.com/s1ghhh/lm-evaluation-harness.git@3196e907fa195b684470a913c7235ed7f08a4383#egg=lm_eval
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.55.3
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.1
pandas==2.2.3
parso==0.8.4
path==17.0.0
pathvalidate==3.2.1
peft==0.9.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.22.0
sacrebleu==2.4.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.20.3
torch==2.5.0
torch-fidelity==0.3.0
torchaudio==2.5.0
torchinfo==1.8.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
trl==0.12.1
typepy==1.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.29.0
wandb==0.18.7
wcwidth==0.2.13
websockets==11.0.3
word2number==1.1
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
zstandard==0.23.0
